% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{et_liste}
\alias{et_liste}
\title{Liste der verfügbaren Ertragstafeln}
\usage{
et_liste()
}
\value{
Knappe Beschreibung der vorhandenen Ertragstafeln in einem Dataframe
mit den Spalten \code{Baumart}, \code{ArtCodeNds}, \code{Tafelname}, \code{Jahr}, \code{Region}.

\tabular{ll}{
\strong{Kürzel} \tab \strong{Beschreibung}\cr
\code{Baumart}    \tab deutscher Name der Baumart\cr
\code{ArtCodeNds} \tab Codenummer der niedersächsischen Landesforsten\cr
\code{Tafelname}  \tab Name der Tafel\cr
\code{Jahr}       \tab Jahr der Veröffentlichung\cr
\code{Region}     \tab Gültigkeitsgebiet der Tafel\cr
}
}
\description{
Liste der verfügbaren Ertragstafeln
}
\examples{
et_liste()
}
\seealso{
\code{\link[=et_info]{et_info()}} für detaillierte Informationen zu einer Tafel.
}
\author{
Robert Nuske
}
